\name{hmcfortran-package}
\alias{hmcfortran-package}
\alias{hmcfortran}
\docType{package}
\title{
\packageTitle{hmcfortran}
}
\description{
\packageDescription{hmcfortran}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hmcfortran}
\packageIndices{hmcfortran}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{hmcfortran}

Maintainer: \packageMaintainer{hmcfortran}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
